resolving bugs
