--- src/utils/types/Product.d.ts
@@ -1,11 +1,0 @@
-export interface IProducts {
-  nextPage:number;
-  currentPage: number;
-  previousPage:number;
-  limit:number;
-  data:[];
-  error?:string
-}
-export interface IProductsState {
-  searchProduct: searchProduct;
-}
